Refactor UI into separate components
